#312 Validate existing files

+++ app/models/PlantsModel.php
@@ -463,8 +463,13 @@ class PlantsModel extends \Asatru\Database\Model {
 
             static::raw('UPDATE `' . self::tableName() . '` SET photo = ?, last_edited_user = ?, last_edited_date = CURRENT_TIMESTAMP WHERE id = ?', [self::PLANT_PLACEHOLDER_FILE, $user->get('id'), $plantId]);
         
-            unlink(public_path() . '/img/' . $item->get('photo'));
-            unlink(public_path() . '/img/' . str_replace('_thumb', '', $item->get('photo')));
+            if (file_exists(public_path() . '/img/' . $item->get('photo'))) {
+                unlink(public_path() . '/img/' . $item->get('photo'));
+            }
+            
+            if (file_exists(public_path() . '/img/' . str_replace('_thumb', '', $item->get('photo')))) {
+                unlink(public_path() . '/img/' . str_replace('_thumb', '', $item->get('photo')));
+            }
 
             LogModel::addLog($user->get('id'), $plantId, 'photo', self::PLANT_PLACEHOLDER_FILE, url('/plants/details/' . $plantId));
